Add retry logic for transient 503 errors in Dataflow job polling tests
